Refactor app with Flask blueprints
