+++ Backend/templates/coming_soon.html
@@ -4,6 +4,6 @@
   <div class="container py-5 text-center">
     <h1 class="mb-4">{{ page_title | default('Feature') }} Coming Soon</h1>
     <p class="lead">This section is under development. Please check back later.</p>
-    <a href="{{ url_for('home') }}" class="btn btn-primary mt-3">Return Home</a>
+    <a href="{{ url_for('main.home') }}" class="btn btn-primary mt-3">Return Home</a>
   </div>
 {% endblock %}
