+++ Backend/templates/forgot_password.html
@@ -17,7 +17,7 @@
     {% endif %}
     {% endwith %}
     <h2 class="login-title">Forgot Password</h2>
-    <form action="{{ url_for('forgot_password') }}" method="post" class="login-form">
+    <form action="{{ url_for('auth.forgot_password') }}" method="post" class="login-form">
       <div class="form-group">
         <label for="email">Email</label>
         <input type="email" id="email" name="email" placeholder="abc@analyticsinstitute.edu.au" required>
