+++ Backend/templates/help.html
@@ -9,7 +9,7 @@
   <p>To access dashboards, register for an account or sign in if you already have one. Managers can view managerial dashboards while stakeholders see stakeholder dashboards.</p>
   <h2 class="mt-4">Frequently Asked Questions</h2>
   <ul>
-    <li><strong>How do I reset my password?</strong> Use the <a href="{{ url_for('forgot_password') }}">Forgot Password</a> link on the sign in page.</li>
+    <li><strong>How do I reset my password?</strong> Use the <a href="{{ url_for('auth.forgot_password') }}">Forgot Password</a> link on the sign in page.</li>
     <li><strong>Who do I contact for access issues?</strong> Send an email to <a href="mailto:info@analyticsinstitute.edu.au">info@analyticsinstitute.edu.au</a>.</li>
     <li><strong>Where can I find more documentation?</strong> Additional resources will be added here as features are released.</li>
   </ul>
@@ -20,6 +20,6 @@
     <li>Phone: (+61) 426 799 276</li>
     <li>Address: Pearl Level 10, 607 Bourke St, Melbourne VIC 3000</li>
   </ul>
-  <a class="btn btn-primary mt-3" href="{{ url_for('home') }}">Return Home</a>
+  <a class="btn btn-primary mt-3" href="{{ url_for('main.home') }}">Return Home</a>
 </div>
 {% endblock %}
