+++ Backend/templates/login.html
@@ -17,7 +17,7 @@
       {% endif %}
       {% endwith %}
       <h2 class="login-title">Sign In</h2>
-      <form action="{{ url_for('login') }}" method="post" class="login-form">
+      <form action="{{ url_for('auth.login') }}" method="post" class="login-form">
         <div class="form-group">
           <label for="email">Email</label>
           <input type="email" id="email" name="email" placeholder="abc@analyticsinstitute.edu.au" required>
@@ -30,7 +30,7 @@
           </div>
         </div>
         <button type="submit" class="btn-primary">Sign In</button>
-        <a href="{{ url_for('forgot_password') }}" class="forgot-password">Forgot password?</a>
+        <a href="{{ url_for('auth.forgot_password') }}" class="forgot-password">Forgot password?</a>
       </form>
     </div>
   </main>
