+++ Backend/templates/reports.html
@@ -10,6 +10,6 @@
     <li>Phone: (+61) 426 799 276</li>
     <li>Address: Pearl Level 10, 607 Bourke St, Melbourne VIC 3000</li>
   </ul>
-  <a class="btn btn-primary mt-3" href="{{ url_for('home') }}">Return Home</a>
+  <a class="btn btn-primary mt-3" href="{{ url_for('main.home') }}">Return Home</a>
 </div>
 {% endblock %}
