+++ Backend/templates/website_landing_page.html
@@ -9,7 +9,7 @@
       <h2>Power BI Dashboard Portal –</h2>
       <h2>Analytics Institute of Australia</h2>
       <p>A centralized, real-time intelligence system for decision-making, insight, and strategy</p>
-      <a href="{{ url_for('dashboard') }}" class="btn-get-started">Get Started</a>
+      <a href="{{ url_for('dashboard.dashboard') }}" class="btn-get-started">Get Started</a>
     </div>
   </section>
   <section class="section-features bg-light">
